enumの定数名被りを解消

+++ Project/Application/Enemy/BossEnemy.cpp
@@ -14,7 +14,7 @@
 using namespace GameEngine;
 
 void BossEnemy::Initialize(const float& stageRadius, EnemyAttackManager* enemyAttackManager, GameEngine::Animator* animator,
-    std::array<std::map<std::string, AnimationData>, enemyAnimationType::MaxCount> animationData, GameEngine::DebugRenderer* debugRenderer) {
+    std::array<std::map<std::string, AnimationData>, static_cast<size_t>(enemyAnimationType::MaxCount)> animationData, GameEngine::DebugRenderer* debugRenderer) {
 
     // 取得
     enemyAttackManager_ = enemyAttackManager;
